+++ Blocks5/src/fire.h
@@ -11,7 +11,7 @@ public:
 	Fire(Level& level, const Vec2i& position);
 	~Fire();
 
-	void onRender(RenderLayer layer, const Vec4d& color);
+	void onRender(RenderLayer layer, const Vec4f& color);
 	void updateSprites();
 	void onUpdate();
